Fix bugs: invert priorities/reinterpret MaxAhead
We had listed the priorities as `P1 | .. | P5`, but we regarded P1 as
the highest priority in the policies. I've renamed them now to `PLowest`
to `PHighest` to avoid all ambiguity.

We now also record MaxInFlight by precedence; see comment of `MaxAhead`
for justification.

This also adds some debugging logging that we can leave in the code.